feat: locally simulate and import whitelist and incentive info from aave-merkl-token-whitelist
